Permit libpq env variable usage
libpq uses a lot of environment variables [1] in order to build its
connection parameters.

The usage of these variables was documented but did not work.
Here we remove the default value from Septentrion: if the values are not
defined then psycopg2/libpq will look at its own environment variables.
The default values for psotgresql settings are now the ones from
psycopg2/libpq.

[1] https://www.postgresql.org/docs/10/libpq-envars.html